Refactored markdown formatting functions into MarkdownFormatter.swift. Builds and runs. Not observing anything broken. Will probably need more polish
